Fix migration foreign key type and enable Dark Matter expedition rewards

+++ database/migrations/2025_11_18_000001_create_planet_relocations_table.php
@@ -13,7 +13,7 @@ return new class extends Migration
     {
         Schema::create('planet_relocations', function (Blueprint $table) {
             $table->id();
-            $table->unsignedBigInteger('planet_id');
+            $table->unsignedInteger('planet_id');
             $table->unsignedBigInteger('user_id');
             $table->integer('from_galaxy');
             $table->integer('from_system');
